feat: Support outputing a patch for failed files
This makes it much easier to fix invalid boilerplate.

It required a significant change to the algorithm, as we need to keep
track of where the existing template is in the file, as well as the
current copyright year. We manually parse comments, but only need to
support the kind seen in boilerplates. We enforce having a newline
before (if applicable) and after the boilerplate, trading flexibility
for consistently and implementation simplicity. I feel the code is
overall cleaner.

This commit implements the base logic, but commits improving the
testing and UX should follow.

Signed-off-by: Vincent de Phily <vdephily@redhat.com>